requirements.txt
flask==2.3.2
flask-cors==3.0.10  
openai==1.3.5
faiss-cpu==1.7.4
python-dotenv==1.0.1
numpy==1.26.4



